+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/orc/OrcFilters.scala
@@ -35,6 +35,8 @@ private[orc] object OrcFilters {
   def createFilter(schema: StructType, filters: Seq[Filter]): Option[SearchArgument] = {
     val dataTypeMap = schema.map(f => f.name -> f.dataType).toMap
 
+    // First, tries to convert each filter individually to see whether it's convertible, and then
+    // collect all convertible ones to build the final `SearchArgument`.
     val convertibleFilters = for {
       filter <- filters
       _ <- buildSearchArgument(dataTypeMap, filter, SearchArgumentFactory.newBuilder())
